+++ src/main/resources/static/templates/index.html
@@ -215,6 +215,7 @@
     <script src="../js/custom.js"></script>
       <script src='../js/layermanager.js'></script>
       <script src='../js/SocketConnect.js'></script>
+      <script src='../js/ManualMove.js'></script>
     <script src="../js/map.js"></script>
   </footer>
